feat: switch to ssr for global avg
